+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * BREAKING CHANGE: The `m` parameter in in `Flay` and `Flayable` has been
   existentialized, to be chosen by the caller.
 
+* Added `Flayable1` and `flay1`.
+
 
 # Version 0.1